fix: refacto
